Added the code we used for the genetic algorithm
Alpha_1.py is the code used to find the largest induced forest in d-degenerate graphs.

Alpha_d.py is the code used to find the largest induced d-degenerate subgraphs in D-degenerate graphs.

sample_batch.sh is a sample batching script. FIR Supercluster uses SLURM as its job scheduler. In order to pass it a job, you must use a batching script. Please feel free to use this one as an example.