Added the following files
Stress_Testing.txt shows an example of alpha_1.py with the parameters (100,100,2,2,100). This code took 5 hours to execute

Stress_Testing_Nonforest.txt shows an example of alpha_d.py with the parameters (30,30,3,2,2,100). This code took 8 hours to execute

Fifteen_Vertices_Test_1.txt and Fifteen_Vertices_Test_3.txt show that the genetic algorithm we used in alpha_1.py was *usually* able to find our bound of 3/5 in the 2v1 case relatively quickly.

Fifteen_Vertices_test_2.txt is an example of how it's not perfect and can sometimes not find what we wanted